// inventario_frontend/src/models/JwtResult.ts
export class JwtResult {
    id: string | undefined = undefined;
    username: string | undefined = undefined;
    mail: string | undefined = undefined;
    name: string | undefined = undefined;
    lastname: string | undefined = undefined;
    token: string | undefined = undefined;
    // codigosMenues : string [];
    // roles!: string [];
}